+++ Sources/Client/Model/Channel/ChannelsQuery.swift
@@ -43,8 +43,8 @@ public struct ChannelsQuery: Encodable {
     ///   - options: a query options (see `QueryOptions`).
     public init(filter: Filter = .none,
                 sort: [Sorting] = [],
-                pagination: Pagination = .channelsPageSize,
-                messagesLimit: Pagination = .messagesPageSize,
+                pagination: Pagination = [.channelsPageSize],
+                messagesLimit: Pagination = [.messagesPageSize],
                 options: QueryOptions = []) {
         self.filter = filter
         self.sort = sort
@@ -64,7 +64,7 @@ public struct ChannelsQuery: Encodable {
             try container.encode(sort, forKey: .sort)
         }
         
-        try container.encode(messagesLimit.limit, forKey: .messagesLimit)
+        try container.encode(messagesLimit.limit ?? 0, forKey: .messagesLimit)
         try options.encode(to: encoder)
         try pagination.encode(to: encoder)
     }