+++ Sources/Core/Presenter/Channel/ChannelPresenter.swift
@@ -91,6 +91,10 @@ public final class ChannelPresenter: Presenter {
     /// Uploader for images and files.
     public private(set) lazy var uploader = Uploader()
     
+    /// It will trigger `channel.stopWatching()` if needed when the presenter was deallocated.
+    /// It's no needed if you will disconnect when the presenter will be deallocated.
+    public var stopWatchingIfNeeded = false
+    
     /// Init a presenter with a given channel.
     ///
     /// - Parameters:
@@ -102,7 +106,7 @@ public final class ChannelPresenter: Presenter {
         channelId = channel.id
         self.parentMessage = parentMessage
         self.queryOptions = queryOptions
-        super.init(pageSize: .messagesPageSize)
+        super.init(pageSize: [.messagesPageSize])
         channelAtomic.set(channel)
     }
     
@@ -117,13 +121,19 @@ public final class ChannelPresenter: Presenter {
         parentMessage = nil
         self.queryOptions = queryOptions
         self.showStatuses = showStatuses
-        super.init(pageSize: .messagesPageSize)
+        super.init(pageSize: [.messagesPageSize])
         parse(response: response)
     }
     
     deinit {
-        if channel.didLoad, Client.shared.isConnected {
-            channel.stopWatching()
+        if stopWatchingIfNeeded, channel.didLoad {
+            let channel = self.channel
+            
+            DispatchQueue.global(qos: .background).asyncAfter(deadline: .now() + 1 + .milliseconds(Int.random(in: 0...3000))) {
+                if Client.shared.isConnected {
+                    channel.stopWatching()
+                }
+            }
         }
     }
 }