+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### 🐞 Fixed
-
+- SPM support [#156](https://github.com/GetStream/stream-chat-swift/issues/156).
 
 # [2.0.1](https://github.com/GetStream/stream-chat-swift/releases/tag/2.0.1)
 _April 3, 2020_
@@ -26,7 +26,7 @@ _April 3, 2020_
 # [2.0.0](https://github.com/GetStream/stream-chat-swift/releases/tag/2.0.0)
 _April 2, 2020_
 
-### SteamChat 2.0 here and it's brand new :sparkles: :rocket:
+### StreamChat 2.0 here and it's brand new :sparkles: :rocket:
 
 We've added/removed/fixed/changed a lot of stuff, so it's fair to say that StreamChat 2.0 is everything you've liked about 1.x and nothing you didn't like about it :smile:
 
@@ -38,6 +38,7 @@ Most importantly:
 - We've updated our dependencies, and removed some.
 
 Aside from those, we fixed tons of bugs and polished our API. Now it should be more Swifty :tada:
+
 We're working hard on migration guide for our 1.x users and will publish it shortly.
 You can check out updated docs [here](http://getstream.io/chat/docs/)